refactor: reorganize driver integration suite test files by category
Rename 34 test files to use consistent category prefixes that align
with the naming conventions used in toasty-core and toasty crate tests.

Categories: batch_, crud_, embed_, field_, infra_, relation_, tx_, type_
