+++ crates/toasty-driver-integration-suite/src/tests.rs
@@ -5,49 +5,49 @@ pub mod batch_query;
 pub mod batch_query_dynamic;
 pub mod batch_rollback;
 pub mod batch_update_delete;
-pub mod belongs_to_configured;
-pub mod belongs_to_one_way;
-pub mod belongs_to_self_referential;
-pub mod bigdecimal;
-pub mod composite_key_pagination;
-pub mod connection_per_clone;
-pub mod create_macro;
-pub mod decimal;
-pub mod default_and_update;
-pub mod embedded_enum_data;
-pub mod embedded_enum_index;
-pub mod embedded_enum_unit;
-pub mod embedded_struct;
-pub mod embedded_struct_index;
+pub mod crud_basic;
+pub mod crud_batch_create;
+pub mod crud_composite_key_in_list;
+pub mod crud_composite_key_pagination;
+pub mod crud_create_macro;
+pub mod crud_driver_ops;
+pub mod crud_model_level_key;
+pub mod crud_option_filter;
+pub mod crud_partitioned;
+pub mod crud_query;
+pub mod crud_sort_limit;
+pub mod embed_enum_data;
+pub mod embed_enum_filter;
+pub mod embed_enum_filter_comparison;
+pub mod embed_enum_filter_variant_field;
+pub mod embed_enum_index;
+pub mod embed_enum_unit;
+pub mod embed_struct;
+pub mod embed_struct_index;
 pub mod field_auto;
 pub mod field_column_name;
 pub mod field_column_type;
-pub mod filter_data_enum;
-pub mod filter_data_enum_variant_field;
-pub mod filter_enum_comparison;
-pub mod has_many_batch_create;
-pub mod has_many_crud_basic;
-pub mod has_many_crud_multi_relations;
-pub mod has_many_filter_on_association;
-pub mod has_many_link_unlink;
-pub mod has_many_n_1;
-pub mod has_many_scoped_query;
-pub mod has_one_crud_basic;
-pub mod jiff;
-pub mod missing_registrations;
-pub mod one_model_batch_create;
-pub mod one_model_composite_key_in_list;
-pub mod one_model_crud;
-pub mod one_model_crud_basic_driver_ops;
-pub mod one_model_model_level_key;
-pub mod one_model_option_filter;
-pub mod one_model_partitioned_crud;
-pub mod one_model_query;
-pub mod one_model_sort_limit;
-pub mod preload;
-pub mod reset_db;
-pub mod serialize;
-pub mod sync_send;
+pub mod field_default_and_update;
+pub mod infra_connection_per_clone;
+pub mod infra_missing_registrations;
+pub mod infra_reset_db;
+pub mod infra_sync_send;
+pub mod relation_belongs_to_configured;
+pub mod relation_belongs_to_one_way;
+pub mod relation_belongs_to_self_referential;
+pub mod relation_has_many_batch_create;
+pub mod relation_has_many_crud;
+pub mod relation_has_many_filter;
+pub mod relation_has_many_link_unlink;
+pub mod relation_has_many_multi;
+pub mod relation_has_many_n_plus_1;
+pub mod relation_has_many_scoped_query;
+pub mod relation_has_one_crud;
+pub mod relation_preload;
 pub mod tx_atomic_stmt;
 pub mod tx_interactive;
-pub mod tys;
+pub mod type_bigdecimal;
+pub mod type_decimal;
+pub mod type_jiff;
+pub mod type_primitives;
+pub mod type_serialize;
