+++ crates/toasty-driver-integration-suite/src/tests/embed_enum_data.rs
