+++ crates/toasty-driver-integration-suite/src/tests/embed_struct_index.rs
