+++ crates/toasty-driver-integration-suite/src/tests/relation_belongs_to_self_referential.rs
